Move code-context to Code folder
