+++ src/components/Code/code-context.js
@@ -1,8 +1,8 @@
 import React, { useContext, useMemo } from 'react';
 import { Language } from '@leafygreen-ui/code';
 import LeafyIcon from '@leafygreen-ui/icon';
-import { TabContext } from './Tabs/tab-context';
-import { getPlaintext } from '../utils/get-plaintext';
+import { TabContext } from '../Tabs/tab-context';
+import { getPlaintext } from '../../utils/get-plaintext';
 
 const defaultContextValue = {
   codeBlockLanguage: null,
